Restructure the File errors in controller
